third_party/blink/web_tests/webaudio/AudioBufferSource/audiobuffersource-loop-comprehensive.html

<!DOCTYPE html>
<html>
  <head>
    <title>audiobuffersource-loop-comprehensive.html</title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script src="../resources/audit-util.js"></script>
    <script src="../resources/audiobuffersource-testing.js"></script>
  </head>
  <body>
    <script id="layout-test-code">
      // The following test cases assume an AudioBuffer of length 8 whose PCM
      // data is a linear ramp, 0, 1, 2, 3,... |description| is optional and
      // will be computed from the other parameters. |offsetFrame| is optional
      // and defaults to 0.
      const tests = [
        {
          description:
              'loop whole buffer by default with loopStart == loopEnd == 0',
          loopStartFrame: 0,
          loopEndFrame: 0,
          renderFrames: 16,
          playbackRate: 1,
          expected: [
            0, 1, 2, 3, 4, 5, 6, 7, 0, 1, 2, 3, 4, 5, 6, 7
          ]
        },

        {
          description: 'loop whole buffer explicitly',
          loopStartFrame: 0,
          loopEndFrame: 8,
          renderFrames: 16,
          playbackRate: 1,
          expected: [
            0, 1, 2, 3, 4, 5, 6, 7, 0, 1, 2, 3, 4, 5, 6, 7
          ]
        },

        {
          description: 'loop from middle to end of buffer',
          loopStartFrame: 4,
          loopEndFrame: 8,
          renderFrames: 16,
          playbackRate: 1,
          expected: [
            0, 1, 2, 3, 4, 5, 6, 7, 4, 5, 6, 7, 4, 5, 6, 7
          ]
        },

        {
          description: 'loop from start to middle of buffer',
          loopStartFrame: 0,
          loopEndFrame: 4,
          renderFrames: 16,
          playbackRate: 1,
          expected: [
            0, 1, 2, 3, 0, 1, 2, 3, 0, 1, 2, 3, 0, 1, 2, 3
          ]
        },

        {
          loopStartFrame: 4,
          loopEndFrame: 6,
          renderFrames: 16,
          playbackRate: 1,
          expected: [
            0, 1, 2, 3, 4, 5, 4, 5, 4, 5, 4, 5, 4, 5, 4, 5
          ]
        },

        {
          loopStartFrame: 3,
          loopEndFrame: 7,
          renderFrames: 16,
          playbackRate: 1,
          expected: [
            0, 1, 2, 3, 4, 5, 6, 3, 4, 5, 6, 3, 4, 5, 6, 3
          ]
        },

        {
          loopStartFrame: 4,
          loopEndFrame: 6,
          renderFrames: 16,
          playbackRate: 0.5,
          expected: [
            0, 0.5, 1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5, 5.5, 4, 4.5, 5, 5.5
          ]
        },

        {
          loopStartFrame: 4,
          loopEndFrame: 6,
          renderFrames: 16,
          playbackRate: 1.5,
          expected: [
            0, 1.5, 3, 4.5, 4, 5.5, 5, 4.5, 4, 5.5, 5, 4.5, 4, 5.5, 5, 4.5
          ]
        },

        // Offset past loop end, so playback starts at loop start
        {
          loopStartFrame: 2,
          loopEndFrame: 5,
          renderFrames: 16,
          playbackRate: 1,
          offsetFrame: 6,
          expected: [
            2, 3, 4, 2, 3, 4, 2, 3, 4, 2, 3, 4, 2, 3, 4, 2
          ]
        },

        // Offset before loop start, so start at offset and continue
        {
          loopStartFrame: 3,
          loopEndFrame: 6,
          renderFrames: 16,
          playbackRate: 1,
          offsetFrame: 1,
          expected: [
            1, 2, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4
          ]
        },

        // Offset between loop start and loop end, so start at offset and
        // continue
        {
          loopStartFrame: 3,
          loopEndFrame: 6,
          renderFrames: 16,
          playbackRate: 1,
          offsetFrame: 4,
          expected: [
            4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4
          ]
        },

        {
          description: 'illegal playbackRate of 47 greater than loop length',
          loopStartFrame: 4,
          loopEndFrame: 6,
          renderFrames: 16,
          playbackRate: 47,
          expected: Array(16).fill(0)
        },

        // Try illegal loop-points - they should be ignored and we'll loop the
        // whole buffer.

        {
          description: 'illegal loop: loopStartFrame > loopEndFrame',
          loopStartFrame: 7,
          loopEndFrame: 3,
          renderFrames: 16,
          playbackRate: 1,
          expectWholeBuffer: true,
          expected: [
            0, 1, 2, 3, 4, 5, 6, 7, 0, 1, 2, 3, 4, 5, 6, 7
          ]
        },

        {
          description: 'illegal loop: loopStartFrame == loopEndFrame',
          loopStartFrame: 3,
          loopEndFrame: 3,
          renderFrames: 16,
          playbackRate: 1,
          expectWholeBuffer: true,
          expected: [
            0, 1, 2, 3, 4, 5, 6, 7, 0, 1, 2, 3, 4, 5, 6, 7
          ]
        },

        {
          description: 'illegal loop: loopStartFrame < 0',
          loopStartFrame: -8,
          loopEndFrame: 3,
          renderFrames: 16,
          playbackRate: 1,
          expectWholeBuffer: true,
          expected: [
            0, 1, 2, 3, 4, 5, 6, 7, 0, 1, 2, 3, 4, 5, 6, 7
          ]
        },

        {
          description: 'illegal loop: loopEndFrame > bufferLength',
          loopStartFrame: 0,
          loopEndFrame: 30000,
          renderFrames: 16,
          playbackRate: 1,
          expectWholeBuffer: true,
          expected: [
            0, 1, 2, 3, 4, 5, 6, 7, 0, 1, 2, 3, 4, 5, 6, 7
          ]
        },

        // Start a loop with a duration longer than the buffer.  The output
        // should be the data from frame 1 to 6, and then looping from 3 to 5
        // until 20 frames have been played.
        {
          description: 'loop from 3 -> 6 with offset 1 for 20 frames',
          loopStartFrame: 3,
          loopEndFrame: 6,
          playbackRate: 1,
          offsetFrame: 1,
          renderFrames: 30,
          durationFrames: 20,
          expected: [
            1, 2, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3,
            4, 5, 3, 4, 5, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
          ]
        },

        // Start a loop with a duration less than the length of the looping
        // frames.  The output should be the data from frame 1 to 3, and then
        // stopping because duration = 3
        {
          description: 'loop from 3 -> 8 with offset 1 for 3 frames',
          loopStartFrame: 3,
          loopEndFrame: 8,
          playbackRate: 1,
          offsetFrame: 1,
          durationFrames: 3,
          renderFrames: 30,
          expected: [
            1, 2, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
            0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
          ]
        },

        // Start a loop with a duration less than the length of the looping
        // frames.  The output should be the data from frame 1 to 3, and then
        // stopping because duration = 3
        {
          description: 'loop from 3 -> 8 with offset 7 for 3 frames',
          loopStartFrame: 3,
          loopEndFrame: 8,
          playbackRate: 1,
          offsetFrame: 7,
          durationFrames: 3,
          renderFrames: 30,
          expected: [
            7, 3, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
            0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
          ]
        }

      ];

      const sampleRate = 32768;
      const bufferFrameLength = 8;
      const testSpacingFrames = 32;
      const testSpacingSeconds = testSpacingFrames / sampleRate;
      const totalRenderLengthFrames = testSpacingFrames * tests.length;

      /**
       * Runs a loop test by scheduling an AudioBufferSourceNode with given
       * params.
       * @param {!BaseAudioContext} context
       * @param {number} testNumber
       * @param {!Object} test
       */
      function runLoopTest(context, testNumber, test) {
        const source = new AudioBufferSourceNode(context);
        source.buffer = buffer;
        source.playbackRate.value = test.playbackRate;
        source.loop = true;

        const invalidLoop =
            test.loopStartFrame < 0 ||
            test.loopEndFrame > buffer.length ||
            (test.loopEndFrame !== 0 &&
             test.loopStartFrame >= test.loopEndFrame);

        if (invalidLoop && !test.expectWholeBuffer) {
          throw new Error(
              `Test ${testNumber}: unexpected invalid loop points ` +
              `(start=${test.loopStartFrame}, end=${test.loopEndFrame})`);
        }

        source.loopStart = test.loopStartFrame / context.sampleRate;
        source.loopEnd = test.loopEndFrame / context.sampleRate;

        const offset = test.offsetFrame
            ? test.offsetFrame / context.sampleRate
            : 0;
        source.connect(context.destination);

        const startTime = testNumber * testSpacingSeconds;

        if (test.durationFrames !== undefined) {
          if (!test.renderFrames) {
            throw new Error(
                `renderFrames is required for test ${testNumber}: ` +
                `${test.description || ''}`);
          }
          if (test.durationFrames > testSpacingFrames ||
              test.durationFrames < 0) {
            throw new Error(
                `Test ${testNumber}: durationFrames (${test.durationFrames}) ` +
                `outside the range [0, ${testSpacingFrames}]`
            );
          }
          source.start(startTime,
                       offset,
                       test.durationFrames / context.sampleRate);
        } else if (test.renderFrames !== undefined) {
          if (test.renderFrames > testSpacingFrames ||
              test.renderFrames < 0) {
            throw new Error(
                `Test ${testNumber}: renderFrames (${test.renderFrames}) ` +
                `outside the range [0, ${testSpacingFrames}]`
            );
          }
          source.start(startTime, offset);
          source.stop(
              startTime +
              test.renderFrames / context.sampleRate);
        } else {
          throw new Error(
              `Test ${testNumber} must specify renderFrames ` +
              `and optionally durationFrames`);
        }
      }

      /**
       * Checks all tests' output in the rendered AudioBuffer against expected
       * data.
       * @param {!AudioBuffer} audioBuffer
       */
      function checkAllTests(audioBuffer) {
        const channelData = audioBuffer.getChannelData(0);

        tests.forEach((test, i) => {
          const startFrame = i * testSpacingFrames;
          const expectedLength = test.expected.length;

          // Extract output segment to check
          const actual = channelData.slice(
              startFrame, startFrame + expectedLength);
          const expected = Float32Array.from(test.expected);

          // Assert length matches
          assert_equals(
              actual.length,
              expectedLength,
              `Test ${i}: length mismatch for “${test.description || ''}”`);

          assert_array_equals(
              actual,
              expected,
              `Test ${i}: rendered samples differ for “${test.description || ''}”`);
        });
      }

      // Main test
      promise_test(async () => {
        const context = new OfflineAudioContext(
          1, totalRenderLengthFrames, sampleRate
        );
        buffer = createTestBuffer(context, bufferFrameLength);

        // Schedule all tests
        tests.forEach((t, idx) => runLoopTest(context, idx, t));

        const renderedBuffer = await context.startRendering();
        checkAllTests(renderedBuffer);
      }, 'AudioBufferSource looping test');
    </script>
  </body>
</html>
